+++ main.go
@@ -3,10 +3,12 @@ package main
 import (
 	"encoding/json"
 	"flag"
+	"fmt"
 	"html/template"
 	"io/ioutil"
 	"log"
 	"net/http"
+	"time"
 )
 
 type (
@@ -26,27 +28,60 @@ type (
 		CDP  *CDP   `json:"cdp"`
 	}
 
+	//EqualizerCDP Structure like the json from cdp-equalizer
+	EqualizerCDP struct {
+		Current  *CDP  `json:"current"`
+		KeyPrices []CDP `json:"KeyPrices"`
+		Up       []CDP `json:"Up"`
+		Down     []CDP `json:"Down"`
+	}
+
 	//HTMLPage page html returned
 	HTMLPage struct {
 		PageTitle string
 		History   []History
 		LastDate  string
+		Equalizer EqualizerCDP
 	}
 )
 
 var (
 	//CDPData for page history
-	CDPData  []History
-	jsonFile *string
+	CDPData []History
+	//EqualizerCDPData data json from the cdp equalizer
+	EqualizerCDPData EqualizerCDP
+	jsonFile         *string
+	jsonCDP          *string
 )
 
 func main() {
 	jsonFile = flag.String("file", "", "History file from your cdp.")
 	host := flag.String("host", "localhost:8888", "Hostname or IP and port to use for the service")
+	jsonCDP = flag.String("cdp", "https://anvenouz.be/", "Json CDP from your cdp-equalizer service (see : https://github.com/efemero/cdp-equalizer)")
+
 	flag.Parse()
 
 	if "" == *jsonFile {
 		log.Fatal("You need to add the history file ! see --help.")
+	}
+
+	if "" != *jsonCDP {
+
+		var myClient = &http.Client{Timeout: 10 * time.Second}
+
+		r, err := myClient.Get(*jsonCDP)
+		if err != nil {
+			fmt.Println(err)
+		}
+		defer r.Body.Close()
+
+		err = json.NewDecoder(r.Body).Decode(&EqualizerCDPData)
+
+		if err != nil {
+			log.Fatal(err)
+		}
+		//fmt.Println(EqualizerCDPData.Current.EthNet)
+		//return
 	}
 
 	data, err := ioutil.ReadFile(*jsonFile)
@@ -74,6 +109,7 @@ func index(w http.ResponseWriter, r *http.Request) {
 		PageTitle: "CDP - HISTORY",
 		History:   CDPData,
 		LastDate:  lastDate,
+		Equalizer: EqualizerCDPData,
 	}
 	err := tmpl.Execute(w, data)
 	if nil != err {