+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## 1.0.0
 
+- Add current CDP information
 - Optimise template
 - Correct max size from cdp data